pyproject.toml
[project]
name = "skreader"
version = "1.0.0"
description = "Python library and command line tool for SEKONIC spectrometers remote control."
authors = [{ name = "Oleksandr Reshetniak" }]
maintainers = []
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
homepage = "https://github.com/akares/skreader-py"
repository = "https://github.com/akares/skreader-py"
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Hardware :: Universal Serial Bus (USB)",
]
include = ["LICENSE.txt", "README.md"]
packages = [{ include = "skreader" }]

[tool.poetry.scripts]
skreader = 'skreader.cli:run'

[tool.poetry.dependencies]
python = ">=3.9,<4"
rich-argparse = ">=1.7.0,<2.0.0"
pyusb = ">=1.2.1"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
flake8 = "^7.0.0"
isort = "^5.12.0"
mypy = ">=1.11.0,<1.13.0"
pytest = "^8.3.5"

[tool.black]
line-length = 80
target-version = ["py39", "py310", "py311", "py312", "py313"]

[tool.pyright]
mode = "strict"
